backend/requirements.txt
fastapi
langchain-openai
loguru
python-dotenv
langchain-core
uvicorn[standard]